feat: disengage `FileServer` from `FileServerHandler`
