+++ src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod error;
+pub mod fmt;
 pub mod signal;
